requirements.txt
streamlit
python-dotenv
google-genai
openai
langchain
langchain_openai